% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvchol.R
\name{mvchol}
\alias{mvchol}
\title{Multivariate Cholesky Block Decomposition}
\usage{
mvchol(S,t,j)
}
\arguments{
\item{S}{Sample covariance matrix of the multivariate longitudinal data.}

\item{t}{positive integer indicating the number of repeated measurements.}

\item{j}{positive integer representing the number of outcomes.}
}
\value{
Elements from MCBD of the covariance matrix:
\itemize{
  \item T.mat is the regression coefificients matrix \eqn{T} matrix in the MCBD.
  \item Phit is \eqn{\Phi_{t}} matrix with regression coefficients stacked as \eqn{\phi_{t1}, \phi_{t2}, ..., \phi_{tk}}.
  \item bigPhi is \eqn{\Phi} matrix with regression coefficients stacked as \eqn{\Phi_{2}, \Phi_{3}, ..., \Phi_{T}}.
  \item D is the innovation variance matrix with block diagonal elements \eqn{D_1, D_2, \ldots , D_t}.
  \item Dt are the j*j innovation variance matrices.
}
}
\description{
mvchol returns modified Cholesky block decomposition (MCBD) of the sample covariance matrix for multivariate longitudinal data with j outcomes.
}
\examples{
data(Tcells)
S <- cov(Tcells)
t <- 10
j <- 4
MCBD <- mvchol(S,t,j)
names(MCBD)
}
\references{
Kohli, P. Garcia, T. and Pourahmadi, M. 2016 Modeling the Cholesky Factors of Covariance Matrices of Multivariate Longitudinal Data, Journal of Multivariate Analysis, 145, 87-100.
}
