% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.R
\name{mvp}
\alias{mvp}
\title{Profile Plot for a Longitudinal Outcome}
\usage{
mvp(data,time,mean=TRUE,title="",xlabel="",ylabel="",scol="gray", mcol="black",plot=TRUE,lwd.mean=2)
}
\arguments{
\item{data}{a data frame (or matrix) with n rows for subjects and T columns for the repeated measurements.}

\item{time}{a vector with T equally or unequally spaced time points.}

\item{mean}{a logical indicating whether average response should be plotted. The default is TRUE.}

\item{title}{a character string indicating title of the profile plot. The default is blank.}

\item{xlabel}{a character string indicating label for the x-axis. The default is blank.}

\item{ylabel}{a character string indicating label for the y-axis. The default is blank.}

\item{scol}{color option for lines representing subjects. The default is gray.}

\item{mcol}{color option for the average response. The default is black.}

\item{plot}{logical indicating whether profile plot is returned or not. The default is TRUE.}

\item{lwd.mean}{integer for line width of the average. The default is 2.}
}
\value{
\itemize{
  \item mean.data is the average response.
  \item profile plot if plot is TRUE.
}
}
\description{
mvp returns profile plot for a longitudinal outcome with repeated measurements. This plot shows the change in variable over the repeated measurements for each subject and the average of all subjects.
}
\examples{
data(Tcells)
time <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
j <- 4
n <- 44
gene.names <- c("FYB", "CD69", "IL2RG", "CDC2")
par(mfrow=c(2,2))
for(i in 1:j){
 mvp(Tcells[,seq(i, ncol(Tcells), j)],time,mean=TRUE,title=gene.names[i],xlabel="Time points",ylabel="Expression Response",scol="gray",mcol="black",plot=TRUE,lwd.mean=2)
}
}
