% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpf.R
\name{mvpf}
\alias{mvpf}
\title{Profile Plot for a Longitudinal Outcome with it's Estimated Mean}
\usage{
mvpf(data,time,mean.fitted,title="",xlabel="",ylabel="",scol="gray", mcol="black",fcol="red",lwd.mean=2,lwd.fit=2,lty.fit=2)
}
\arguments{
\item{data}{a data frame (or matrix) with n rows for subjects and T columns for the repeated measurements.}

\item{time}{vector with T equally or unequally spaced time points.}

\item{mean.fitted}{vector with estimated average values.}

\item{title}{a character string indicating title of the profile plot. The default is blank.}

\item{xlabel}{a character string indicating label for the x-axis. The default is blank.}

\item{ylabel}{a character string indicating label for the y-axis. The default is blank.}

\item{scol}{color option for lines representing subjects. The default is gray.}

\item{mcol}{color option for the average response. The default is black.}

\item{fcol}{color option for the estimated average response. The default is red.}

\item{lwd.mean}{integer for line width of the average. The default is 2.}

\item{lwd.fit}{integer for line width of the estimated average. The default is 2.}

\item{lty.fit}{integer for line width of the estimated average. The default is 2.

\itemize{
  \item profile plot with observed and estimated mean values.
}}
}
\description{
mvpf returns profile plot for longitudinal outcome. It includes the observed and estimated mean values.
}
\examples{
data(Tcells)
time <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
j <- 4
n <- 44
gene.names <- c("FYB", "CD69", "IL2RG", "CDC2")
par(mfrow=c(2,2))
for(i in 1:j){
data.gene <- Tcells[,seq(i, ncol(Tcells), j)]
mean.gene <- apply(data.gene,2,mean)
 mvpf(data.gene,time,mean.gene,title=gene.names[i],xlabel="Hours",ylabel="Expression Response",scol="gray", mcol="black",fcol="red",lwd.mean=2,lwd.fit = 3,lty.fit =2)
}
}
