% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvr.R
\name{mvr}
\alias{mvr}
\title{Multivariate Regressograms}
\usage{
mvr(data, time, j, N, inno=FALSE, inverse=FALSE, loginno=TRUE, plot=TRUE, pch.plot=19, par1.r=2, par2.r=2, par1.d=2, par2.d=2)
}
\arguments{
\item{data}{a data frame (or matrix) with n rows for subjects and T columns for the repeated measurements.}

\item{time}{a vector with T equally or unequally spaced time points.}

\item{j}{a positive integer for the number of outcomes.}

\item{N}{a positive integer for the number of subjects.}

\item{inno}{a logical indicating if elements of innovation variance matrices be used innovariogram plot. The default is FALSE.}

\item{inverse}{a logical indicating if elements of inverse innovation variance matrices be used innovariogram plot. The default is FALSE.}

\item{loginno}{a logical indicating if elements of log innovation variance matrices be used innovariogram plot. The default is TRUE.}

\item{plot}{logical indicating whether multivariate regressograms are returned or not. The default is TRUE.}

\item{pch.plot}{a integer indicating type of symbols to be used in  multivariate regressograms. The default is 19 for a solid dot.}

\item{par1.r}{a positive integer indicating number of rows in multiple regressogram plots. The default is 2.}

\item{par2.r}{a positive integer indicating number of columns in multiple regressogram plots. The default is 2.}

\item{par1.d}{a positive integer indicating number of rows in multiple innovariogram plots. The default is 2.}

\item{par2.d}{a positive integer indicating number of columns in multiple innovariogram plots. The default is 2.}
}
\value{
Multivariate regressograms are returned and following elements of modified Cholesky block decomposition:
\itemize{
 \item Phit are the correlation coefficient matrices obtained from sample covariance matrix.
 \item Phi.plot the elements from jXj Phit matrices obtained from sample covariance matrix. This is in a list format where each element of the list represents elements for each regressogram plot.
 \item D.elements are the innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are not plotted by default.
 \item Dinv.elements are the inverse innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are not plotted by default.
 \item logD.elements are the log innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are plotted by default.
}
}
\description{
mvr returns multivariate regressograms for multivariate longitudinal data with j outcomes.
}
\examples{
data(Tcells)
time <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
j <- 4
n <- 44
MVR <- mvr(Tcells,time,j,n,inno=FALSE,inverse=FALSE,loginno=TRUE,plot=TRUE,pch.plot=19,par1.r = 2,par2.r = 2,par1.d=2,par2.d=2)
}
\references{
Kohli, P. Garcia, T. and Pourahmadi, M. 2016 Modeling the Cholesky Factors of Covariance Matrices of Multivariate Longitudinal Data, Journal of Multivariate Analysis, 145, 87-100.

Kohli, P. Du, X. and Shen, H. 2020+ Multivariate Longitudinal Graphical Models (MLGM): An R Package for Visualizing and Modeling Mean and Dependence Patterns in Multivariate Longitudinal Data, submitted.
}
