% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrf.R
\name{mvrf}
\alias{mvrf}
\title{Multivariate Regressograms with Estimated Values}
\usage{
mvrf(data,time,j,N,inno=FALSE,inverse=FALSE,loginno=TRUE,sigma.fitted,mcol="black",fcol="red",pch.plot=19,par1.r=2,par2.r=2,par1.d=2,par2.d=2,lwd.fit=2,lty.fit=1)
}
\arguments{
\item{data}{a data frame (or matrix) with n rows for subjects and T columns for the repeated measurements.}

\item{time}{vector with T equally or unequally spaced time points.}

\item{j}{a positive integer for the number of outcomes.}

\item{N}{a positive integer for the number of subjects.}

\item{inno}{a logical indicating if elements of innovation variance matrices be used innovariogram plot. The default is FALSE.}

\item{inverse}{a logical indicating if elements of inverse innovation variance matrices be used innovariogram plot. The default is FALSE.}

\item{loginno}{a logical indicating if elements of log innovation variance matrices be used innovariogram plot. The default is TRUE.}

\item{sigma.fitted}{Tj X Tj, positive definite estimate of the covariance matix for multivariate longitudinal data with j outcomes.}

\item{mcol}{color option for the average response. The default is black.}

\item{fcol}{color option for the estimated average response. The default is red.}

\item{pch.plot}{a integer indicating type of symbols to be used in  multivariate regressograms. The default is 19 for a solid dot.}

\item{par1.r}{a positive integer indicating number of rows in multiple regressogram plots. The default is 2.}

\item{par2.r}{a positive integer indicating number of columns in multiple regressogram plots. The default is 2.}

\item{par1.d}{a positive integer indicating number of rows in multiple innovariogram plots. The default is 2.}

\item{par2.d}{a positive integer indicating number of columns in multiple innovariogram plots. The default is 2.}

\item{lwd.fit}{integer for line width of the estimated values. The default is 2.}

\item{lty.fit}{integer for line width of the estimated values. The default is 2.}
}
\value{
Multivariate regressograms with fitted values are returned and following elements of modified Cholesky block decomposition:
\itemize{
 \item Phit are the correlation coefficient matrices obtained from sample covariance matrix.
 \item Phi.plot the elements from jXj Phit matrices obtained from sample covariance matrix. This is in a list format where each element of the list represents elements for each regressogram plot.
 \item Phi.fitted.plot the elements from jXj Phit matrices obtained from estimated covariance matrix. This is in a list format where each element of the list represents the estimated elements for each regressogram plot.
 \item D.elements are the innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are not plotted by default.
 \item D.fitted.elements are the innovation variance matrices obtained from estimated covariance matrix. This is in matrix format where each row represents the estimated elements for each innovariogram plot. These elements are not plotted by default.
 \item Dinv.elements are the inverse innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are not plotted by default.
 \item Dinv.fitted.elements are the inverse innovation variance matrices obtained from estimated covariance matrix. This is in matrix format where each row represents the estimated elements for each innovariogram plot. These elements are not plotted by default.
 \item logD.elements are the log innovation variance matrices obtained from sample covariance matrix. This is in matrix format where each row represents the elements for each innovariogram plot. These elements are plotted by default.
 \item logD.fitted.elements are the log innovation variance matrices obtained from estimated covariance matrix. This is in matrix format where each row represents the estimated elements for each innovariogram plot. These elements are plotted by default.
}
}
\description{
mvrf returns multivariate regressograms for multivariate longitudinal data with j outcomes. It includes the estimated values for the multivariate regressograms.
}
\examples{
data(Tcells)
time <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
j <- 4
N <- 44
data(Cov_est)
sigma.fitted <- Cov_est
MVRF <- mvrf(Tcells,time,j,N,inno=FALSE,inverse=FALSE,loginno=TRUE,as.matrix(Cov_est),mcol="black",fcol="red",pch.plot=19,par1.r=1,par2.r=1,par1.d=1,par2.d=1,lwd.fit=2,lty.fit=1)
}
\references{
Kohli, P. Garcia, T. and Pourahmadi, M. 2016 Modeling the Cholesky Factors of Covariance Matrices of Multivariate Longitudinal Data, Journal of Multivariate Analysis, 145, 87-100.

Kohli, P. Du, X. and Shen, H. 2020+ Multivariate Longitudinal Graphical Models (MLGM): An R Package for Visualizing and Modeling Mean \& Dependence Patterns in Multivariate Longitudinal Data, submitted.
}
