% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrmat.R
\name{mvrmat}
\alias{mvrmat}
\title{Design Matrices for Modeling Elements of Regressograms}
\usage{
mvrmat(time,j,r,s)
}
\arguments{
\item{time}{a vector with T equally or unequally spaced time points.}

\item{j}{a positive integer for the number of outcomes.}

\item{r}{a positive integer for the order of time-lag polynomial in constructing design matrix for regressogram.}

\item{s}{a positive integer for the order of time polynomial in constructing design matrix for innovariogram.}
}
\value{
design matrices for models selected for regressogram and innovariogram. Specifically,

\itemize{
  \item U is the design matrix for time-lag polynomials selected for regressogram.
  \item V is the design matrix for time polynomials selected for log innovariogram.
}
}
\description{
mvrmat returns design matrices for polynomial models used for modeling elements of modified Cholesky block decomposition.
}
\examples{
data(Tcells)
time <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
t <- length(time)
j <- 4
n <- 44
r <- 3
s <- 2
design_mat <- mvrmat(time,j,r,s)
}
\references{
Kohli, P. Garcia, T. and Pourahmadi, M. 2016 Modeling the Cholesky Factors of Covariance Matrices of Multivariate Longitudinal Data, Journal of Multivariate Analysis, 145, 87-100.
}
