% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbc2}
\alias{pbc2}
\title{Primary Biliary Cirrhosis for placebo and drug D-pencillamine (DPCA)}
\format{
A data frame with 82 rows and 21 columns.
\describe{
\item{rows}{first 40 rows are for the patients in placebo group and next 42 for those in DPCA group.}
\item{sr1}{serum bilirubin  measurements in mg/dl at the first visit.}
\item{al1}{serum albumin measurements in mg/dl at the first visit.}
\item{pr1}{prothrombin measurementrs in seconds at the first visit.}
\item{sr2}{serum bilirubin  measurements in mg/dl at the second visit.}
\item{al2}{serum albumin measurements in mg/dl at the second visit.}
\item{pr2}{prothrombin measurementrs in seconds at the second visit.}
\item{sr3}{serum bilirubin  measurements in mg/dl at the third visit.}
\item{al3}{serum albumin measurements in mg/dl at the third visit.}
\item{pr3}{prothrombin measurementrs in seconds at the third visit.}
\item{sr4}{serum bilirubin  measurements in mg/dl at the fourth visit.}
\item{al4}{serum albumin measurements in mg/dl at the fourth visit.}
\item{pr4}{prothrombin measurementrs in seconds at the fourth visit.}
\item{sr5}{serum bilirubin  measurements in mg/dl at the fifth visit.}
\item{al5}{serum albumin measurements in mg/dl at the fifth visit.}
\item{pr5}{prothrombin measurementrs in seconds at the fifth visit.}
\item{sr6}{serum bilirubin  measurements in mg/dl at the sixth visit.}
\item{al6}{serum albumin measurements in mg/dl at the sixth visit.}
\item{pr6}{prothrombin measurementrs in seconds at the sixth visit.}
\item{sr7}{serum bilirubin  measurements in mg/dl at the seventh visit.}
\item{al7}{serum albumin measurements in mg/dl at the seventh visit.}
\item{pr7}{prothrombin measurementrs in seconds at the seventh visit.}
}
}
\source{
{R package joineRML} <https://cran.r-project.org/web/packages/joineRML/index.html>
}
\usage{
pbc2
}
\description{
A data set with three biomarkers recorded at seven time points for 40 patients in the placebo group and 42 in the DPCA group.
}
\keyword{datasets}
