% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tcells}
\alias{Tcells}
\title{Microarray T-cell acitivation}
\format{
A data frame with 44 rows and 40 columns.
\describe{
\item{rows}{independent and identical replications.}
\item{FYB.0}{expression levels for gene FYB at the time of treatment.}
\item{CD69.0}{expression levels for CD 69 at the time of treatment.}
\item{IL2RG.0}{expression levels for interleuking gene IL-2Rgamma at the time of treatment.}
\item{CDC2.0}{expression levels for cyclin A2 gene at the time of treatment.}
\item{FYB.2}{expression levels for gene FYB 2 hours after treatment.}
\item{CD69.2}{expression levels for CD 69 2 hours after treatment.}
\item{IL2RG.2}{expression levels for interleuking gene IL-2Rgamma 2 hours after treatment.}
\item{CDC2.2}{expression levels for cyclin A2 gene 2 hours after treatment.}
\item{FYB.4}{expression levels for gene FYB 4 hours after treatment.}
\item{CD69.4}{expression levels for CD 69 4 hours after treatment.}
\item{IL2RG.4}{expression levels for interleuking gene IL-2Rgamma 4 hours after treatment.}
\item{CDC2.4}{expression levels for cyclin A2 gene 4 hours after treatment.}
\item{FYB.6}{expression levels for gene FYB 6 hours after treatment.}
\item{CD69.6}{expression levels for CD 69 6 hours after treatment.}
\item{IL2RG.6}{expression levels for interleuking gene IL-2Rgamma 6 hours after treatment.}
\item{CDC2.6}{expression levels for cyclin A2 gene 6 hours after treatment.}
\item{FYB.8}{expression levels for gene FYB 8 hours after treatment.}
\item{CD69.8}{expression levels for CD 69 8 hours after treatment.}
\item{IL2RG.8}{expression levels for interleuking gene IL-2Rgamma 8 hours after treatmentt.}
\item{CDC2.8}{expression levels for cyclin A2 gene 8 hours after treatment.}
\item{FYB.18}{expression levels for gene FYB 18 hours after treatment.}
\item{CD69.18}{expression levels for CD 69 18 hours after treatment.}
\item{IL2RG.18}{expression levels for interleuking gene IL-2Rgamma 18 hours after treatment.}
\item{CDC2.18}{expression levels for cyclin A2 gene 18 hours after treatment.}
\item{FYB.24}{expression levels for gene FYB 24 hours after treatment.}
\item{CD69.24}{expression levels for CD 69 24 hours after treatment.}
\item{IL2RG.24}{expression levels for interleuking gene IL-2Rgamma 24 hours after treatment.}
\item{CDC2.24}{expression levels for cyclin A2 gene 24 hours after treatment.}
\item{FYB.32}{expression levels for gene FYB 32 hours after treatment.}
\item{CD69.32}{expression levels for CD 69 32 hours after treatment.}
\item{IL2RG.32}{expression levels for interleuking gene IL-2Rgamma 32 hours after treatment.}
\item{CDC2.32}{expression levels for cyclin A2 gene 32 hours after treatment.}
\item{FYB.48}{expression levels for gene FYB 48 hours after treatment.}
\item{CD69.48}{expression levels for CD 69 48 hours after treatment.}
\item{IL2RG.48}{expression levels for interleuking gene IL-2Rgamma 48 hours after treatment.}
\item{CDC2.48}{expression levels for cyclin A2 gene 48 hours after treatment.}
\item{FYB.72}{expression levels for gene FYB 72 hours after treatment.}
\item{CD69.72}{expression levels for CD 69 72 hours after treatment.}
\item{IL2RG.72}{expression levels for interleuking gene IL-2Rgamma 72 hours after treatment.}
\item{CDC2.72}{expression levels for cyclin A2 gene 72 hours after treatment.}
}
}
\source{
{R package longitudinal}<https://cran.r-project.org/web/packages/longitudinal/>
}
\usage{
Tcells
}
\description{
A data set with 44 replications of four genes involved in T-cells activation. For each gene there are ten repeated measurements recorded at 0, 2, 4, 6, 8, 18, 24, 32, 48, 72 hours after the treatement.
}
\keyword{datasets}
